Fix a couple of warnings.

+++ n2o/tap.hpp
@@ -74,7 +74,7 @@ private:
     call(char const* symbol, std::vector<v8::Handle<v8::Value> > const& args = arguments_t()) {
         v8::Local<v8::Object> obj = v8::Local<v8::Object>::New(v8::Isolate::GetCurrent(), t_);
         v8::Local<v8::String> sym = string(symbol).As<v8::String>();
-        node::MakeCallback(v8::Isolate::GetCurrent(), obj, sym, args.size(), 
+        node::MakeCallback(v8::Isolate::GetCurrent(), obj, sym, static_cast<int>(args.size()),
                 const_cast<v8::Handle<v8::Value>*>(&*args.begin())); // XXX const correctness in node
     }
